feat: dev-3 Refactor of files naming.
