+++ internal/domain/agg/msg.go
