+++ internal/domain/agg/optsableQueryInterface.go
